Progress update on vehicle modularization
The code has been reorganized into core, vehicles, and plugins folders.
Vehicle objects have been setup as their own vehicle definition files.
Most methods have now been integrated into the vehicle class. The
google maps stuff hasn’t been activated yet but has been started in the
class. A number of things in the updater and destructor are commented
out for the maps. So the remaining work is running maps with the
vehicle objects. They are working fine in the sidebar tables and are
being destructed as much as we can manage to do with delete being
deprecated.