Connected table hover and click to aircraft icons
CSS - updated the vehicle table rows and pointer icon to react more as
expected.

Created select, unselect, and hover functions for the vehicle icons
which enlarge and change the icon color. The hover function just
changes the icon color. When the table row is clicked (and info table
opened) the icon also enlarges.