Connecting marker objects and info windows to the vehicle data tables
Hovering on and clicking on vehicle markers updates the vehicle tables.
Remaining issue is coordinating the marker inforwindow with the table.
These can easily get out of sync, will need to add more checks on
whether the info window is open or the table detail is shown to remain
in sync.